+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [7.1.0] - 2019-10-09
+
 ### Added
 - Listing & enumerating webhooks
 - Creating, updating, and deleting webhooks